Split entryState specs
Split by topic.
